# doc/openapi.yaml
openapi: 3.0.3
info:
  title: HealthPorta Healthcare MRF API
  version: "1.0.0"
  description: |
    OpenAPI description for the HealthPorta Healthcare MRF API.  The API
    is versioned under `/api/v1` and served by a Sanic application.  All
    endpoints return JSON.  Unless noted otherwise, error responses use
    the `Error` schema.
servers:
  - url: http://localhost:8085/api/v1
paths:
...
  /geo/get:
    get:
      summary: Echo geo metadata for quick status checks
      parameters:
        - name: zip_code
          in: query
          required: true
          schema:
            type: number
          description: ZIP code provided as a numeric value
        - name: lat
          in: query
          required: true
          schema:
            type: number
          description: Latitude to report back in the response metadata
      responses:
        '200':
          description: Minimal geo response with release metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoPingResponse'
  /geo/zip/{zip_code}:
    get:
      summary: Look up ZIP centroid and state
      parameters:
        - name: zip_code
          in: path
          required: true
          schema:
            type: string
          description: 5-digit ZIP code (padding handled server-side)
      responses:
        '200':
          description: ZIP geolocation payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoZipResponse'
        '404':
          description: ZIP code not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Underlying TIGER tables unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /geo/city:
    get:
      summary: List ZIP codes for a given city and optional state
      parameters:
        - name: city
          in: query
          required: true
          schema:
            type: string
          description: City name to search (case insensitive)
        - name: state
          in: query
          required: false
          schema:
            type: string
            maxLength: 2
          description: Optional 2-letter state abbreviation to narrow matches
      responses:
        '200':
          description: Matching ZIP codes for the requested city/state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoCityResponse'
        '404':
          description: No ZIPs found for the requested filters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /geo/states:
    get:
      summary: Aggregated ZIP and city counts per state
      parameters:
        - name: sort
          in: query
          schema:
            type: string
            enum: [population, zip_count, state]
          description: Field used to order the states (defaults to population)
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
          description: Sort direction (defaults to desc)
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Optional maximum number of rows to return
        - name: top_zip_limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 10
          description: Number of top ZIP codes (by population) to include per state (default 5)
      responses:
        '200':
          description: Aggregated state-level geo metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoStateListResponse'
        '400':
          description: Invalid sort/order parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /geo/state/{state}/cities:
    get:
      summary: Top cities in a state by population (based on local geo data)
      parameters:
        - name: state
          in: path
          required: true
          schema:
            type: string
            maxLength: 2
          description: Two-letter state abbreviation
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
          description: Number of cities to include (defaults to 20, max 200)
      responses:
        '200':
          description: Ordered list of cities with population/ZIP counts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoStateCitiesResponse'
        '404':
          description: No city data found for the requested state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      additionalProperties: true
...
    GeoPingResponse:
      type: object
      properties:
        date:
          type: string
          format: date-time
        release:
          type: string
        environment:
          type: string
      additionalProperties: true
    GeoZipResponse:
      type: object
      properties:
        zip_code:
          type: string
        lat:
          type: number
        long:
          type: number
        state:
          type: string
        city:
          type: string
        state_name:
          type: string
        county_name:
          type: string
        timezone:
          type: string
      additionalProperties: false
    GeoCityResponse:
      type: object
      properties:
        normalized_city:
          type: string
        state:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              zip_code:
                type: string
              state:
                type: string
              state_name:
                type: string
              city:
                type: string
              county_name:
                type: string
              lat:
                type: number
              long:
                type: number
              timezone:
                type: string
            additionalProperties: false
      additionalProperties: false
    GeoTopZip:
      type: object
      properties:
        zip_code:
          type: string
        city:
          type: string
        population:
          type: integer
        lat:
          type: number
        long:
          type: number
      required: [zip_code, city, population]
      additionalProperties: false
    GeoStateSummary:
      type: object
      properties:
        state:
          type: string
        state_name:
          type: string
          nullable: true
        zip_count:
          type: integer
        city_count:
          type: integer
        population:
          type: integer
        avg_lat:
          type: number
          nullable: true
        avg_long:
          type: number
          nullable: true
        top_zips:
          type: array
          items:
            $ref: '#/components/schemas/GeoTopZip'
      required: [state, zip_count, city_count, population, top_zips]
      additionalProperties: false
    GeoStateListResponse:
      type: object
      properties:
        generated:
          type: string
          format: date-time
        states:
          type: array
          items:
            $ref: '#/components/schemas/GeoStateSummary'
      required: [states]
      additionalProperties: false
    GeoStateCitiesResponse:
      type: object
      properties:
        state:
          type: string
        limit:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              city:
                type: string
              state:
                type: string
              zip_count:
                type: integer
              population:
                type: integer
              avg_lat:
                type: number
              avg_long:
                type: number
            additionalProperties: false
      additionalProperties: false
...